breaks out prepend and hooks
